+++ app/src/main/java/com/example/baseproject/ui/playlist/customplaylist/CustomPLaylistDialog.kt
@@ -21,7 +21,7 @@ import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
 class CustomPLaylistDialog(
     allPlaylist: MutableList<LibraryItem>,
-    private var songId: Int
+    private var songId: Long
 ) :
     BottomSheetDialogFragment(),
     OnItemClickListener {
@@ -35,7 +35,6 @@ class CustomPLaylistDialog(
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
         val dialog = BottomSheetDialog(requireContext(), theme)
         dialog.setOnShowListener {
-
             val bottomSheetDialog = it as BottomSheetDialog
             val parentLayout =
                 bottomSheetDialog.findViewById<View>(com.google.android.material.R.id.design_bottom_sheet)
